+++ trace-cobertura/src/main/java/se/de/hu_berlin/informatik/spectra/provider/tracecobertura/infrastructure/CompressedTrace.java
@@ -56,7 +56,7 @@ public class CompressedTrace extends CompressedTraceBase<int[],IntArrayWrapper> 
 	}
 	
 	@Override
-	public int getMaxStoredValue() {
+	public long getMaxStoredValue() {
 		if (getChild() == null) {
 			int max = 0;
 			for (int[] i : getCompressedTrace()) {
@@ -66,7 +66,7 @@ public class CompressedTrace extends CompressedTraceBase<int[],IntArrayWrapper> 
 			}
 			return max;
 		} else {
-			int max = getChild().getMaxStoredValue();
+			long max = getChild().getMaxStoredValue();
 			Iterator<Entry<Integer, int[]>> entrySetIterator = getRepetitionMarkers().entrySetIterator();
 			while (entrySetIterator.hasNext()) {
 				Entry<Integer, int[]> entry = entrySetIterator.next();
